[None][chore] Relocate warmup_heuristic_topk_decode to dsa.py (#13477 review follow-up)
PR #13477 reviewers (hyukn, yuxianq) requested on 2026-04-30 that
`warmup_heuristic_topk_decode` and its module-level idempotency cache
live next to their sole caller in
`tensorrt_llm/_torch/attention_backend/sparse/dsa.py` rather than in
`tensorrt_llm/_torch/custom_ops/cpp_custom_ops.py` (which is the wrong
abstraction layer — it hosts `torch.library` op registrations, not
caller-specific warmup helpers). The relocation was acknowledged at
merge time but deferred to a follow-up; this commit closes that loop.

Mechanical move of:
  - `_HEURISTIC_TOPK_WARMUP_DONE` (Set) + `_HEURISTIC_TOPK_WARMUP_LOCK`
    (threading.Lock) idempotency state
  - `def warmup_heuristic_topk_decode(...)` (52 lines, including
    docstring; behavior unchanged)
  - `import threading` and `Set` from `typing` (dropped from
    cpp_custom_ops.py since no other code in that file uses them)

Caller in `Indexer.setup_module()` (dsa.py:1306) is updated from a
conditional cross-module import + `cpp_custom_ops.warmup_heuristic_topk_decode(...)`
to a direct local call. No other call sites exist in the tree.

This is a no-op for behavior: the function body, idempotency keying
(device, top_k, hint_size, num_cols), and CUDA Graph capture ordering
are byte-identical to the merged version from PR #13477.
