+++ tensorrt_llm/_torch/attention_backend/sparse/dsa.py
@@ -1,8 +1,9 @@
 """Dense Sparse Attention (DSA) backend for TRT-LLM with indexer-based TopK selection."""
 import math
+import threading
 from contextlib import contextmanager
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, List, Optional, Tuple, Union
+from typing import TYPE_CHECKING, List, Optional, Set, Tuple, Union
 
 import torch
 import torch.nn as nn
@@ -49,6 +50,58 @@ try:
 except ImportError:
     hadamard_transform = None
     HAS_FAST_HADAMARD = False
+
+# Idempotency guard for warmup_heuristic_topk_decode — keyed by
+# (device_index, top_k, hint_size, num_cols). Prevents repeated allocations
+# and synchronizations when multiple Indexer modules invoke the warmup with
+# the same parameters during model construction.
+_HEURISTIC_TOPK_WARMUP_DONE: Set[Tuple[int, int, int, int]] = set()
+_HEURISTIC_TOPK_WARMUP_LOCK = threading.Lock()
+
+
+def warmup_heuristic_topk_decode(top_k: int = 2048,
+                                 hint_size: int = 2048,
+                                 num_cols: int = 4096) -> None:
+    """Pre-initialize cached hardware attributes in the C++ Scheme X dispatcher.
+
+    The dispatcher inside ``invokeIndexerTopKDecode`` lazily queries
+    ``cudaDeviceGetAttribute`` for ``MultiProcessorCount`` and
+    ``L2CacheSize`` on its first call. Those host-side queries must not
+    be issued during ``cudaStreamBeginCapture / EndCapture``: the values
+    captured there become frozen into the graph and cannot be refreshed
+    across replays on a different device.
+
+    This warmup issues one small heuristic decode call so the static
+    caches are populated before any CUDA Graph capture begins. Must be
+    called from the Indexer setup hook (``layer_idx == 0``) when
+    ``enable_heuristic_topk`` is true.
+
+    Repeated invocations with the same ``(device, top_k, hint_size,
+    num_cols)`` key are short-circuited so that constructing many Indexer
+    modules in the same process does not re-allocate scratch tensors or
+    issue redundant synchronizations.
+    """
+    key = (torch.cuda.current_device(), top_k, hint_size, num_cols)
+    with _HEURISTIC_TOPK_WARMUP_LOCK:
+        if key in _HEURISTIC_TOPK_WARMUP_DONE:
+            return
+        _HEURISTIC_TOPK_WARMUP_DONE.add(key)
+
+    device = torch.device("cuda")
+    logits = torch.zeros((1, num_cols), dtype=torch.float32, device=device)
+    seq_lens = torch.tensor([num_cols], dtype=torch.int32, device=device)
+    indices = torch.empty((1, top_k), dtype=torch.int32, device=device)
+    pre_idx = torch.zeros((1, hint_size), dtype=torch.int32, device=device)
+    scratch = torch.empty((top_k, ), dtype=torch.float32, device=device)
+    torch.ops.trtllm.indexer_topk_decode(logits,
+                                         seq_lens,
+                                         indices,
+                                         1,
+                                         top_k,
+                                         pre_idx=pre_idx,
+                                         heuristic_scratch=scratch)
+    torch.cuda.synchronize()
+
 
 # `block_kv` arg passed to DeepGEMM's `get_paged_mqa_logits_metadata`. This is
 # a SCHEDULE-granularity parameter, not the cache page size. The DG metadata
@@ -1249,8 +1302,7 @@ class Indexer(nn.Module):
             # Populate static caches (sm_count, L2 cache size) inside the C++
             # Scheme X dispatcher before any CUDA Graph capture so the host
             # attribute queries do not end up frozen into a captured graph.
-            from tensorrt_llm._torch.custom_ops import cpp_custom_ops
-            cpp_custom_ops.warmup_heuristic_topk_decode(top_k=self.index_topk)
+            warmup_heuristic_topk_decode(top_k=self.index_topk)
 
     def post_load_weights(self):
         """Fuse wk + weights_proj into single FP32 weight for F.linear GEMM under allow_tf32 (TF32 tensor cores on Ampere+)."""
